requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
Authlib==1.3.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.4
dataclasses-json==0.6.4
distro==1.9.0
exceptiongroup==1.2.0
fastapi==0.109.2
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.42
greenlet==3.0.3
grpcio==1.62.0
grpcio-health-checking==1.62.0
grpcio-tools==1.62.0
h11==0.14.0
httpcore==1.0.4
httpx==0.26.0
httpx-sse==0.4.0
idna==3.6
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.8
langchain-cli==0.0.21
langchain-community==0.0.21
langchain-core==0.1.25
langserve==0.0.42
langsmith==0.1.5
markdown-it-py==3.0.0
marshmallow==3.20.2
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.24.4
openai==1.12.0
orjson==3.9.14
outcome==1.3.0.post0
packaging==23.2
pinecone-client==3.0.3
protobuf==4.25.3
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
Pygments==2.17.2
PySocks==1.7.1
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.0
selenium==4.18.1
shellingham==1.5.4
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.27
sse-starlette==1.8.2
starlette==0.36.3
tenacity==8.2.3
tomlkit==0.12.3
tqdm==4.66.2
trio==0.24.0
trio-websocket==0.11.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.2.1
uvicorn==0.23.2
validators==0.22.0
weaviate-client==4.4.4
webdriver-manager==4.0.1
wsproto==1.2.0
yarl==1.9.4
